+++ NearestAeroplane/src/com/welle/driver/Driver.java
@@ -4,7 +4,7 @@ public interface Driver {
 
 	public void getInputAndStartLoop(long timePeriod);
 
-	public void loopFetchAndCalculate(long timePeriod, Double myLatitude, Double myLongitude);
+	public void loopFetchAndCalculate(long timePeriod, Float myLatitude, Float myLongitude);
 
 	public void stopFetchAndCalculate();
